MainActivityをUserSearchActivityにrenameし、user_search packageに移動
