+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.GithubClient"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".user_search.UserSearchActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.GithubClient">
