+++ app/src/main/java/dev/dai/githubclient/user_search/UserSearchActivity.kt
@@ -1,0 +1,27 @@
+package dev.dai.githubclient.user_search
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.Modifier
+import dev.dai.githubclient.ui.theme.GithubClientTheme
+
+class UserSearchActivity : ComponentActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContent {
+      GithubClientTheme {
+        // A surface container using the 'background' color from the theme
+        Surface(
+          modifier = Modifier.fillMaxSize(),
+          color = MaterialTheme.colors.background
+        ) {
+          UserSearchScreen()
+        }
+      }
+    }
+  }
+}
